Implement Date::operator= inline
operator= can be virtual (and we want this, so we can assign a Date&
to a DerivedDate) but it also needs to be defined. Leaving it pure
virtual will cause linker errors, and it will always be overridden
anyway since the compiler will automatically generate a copy assignment
operator if none is defined.